% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeqr.R
\name{diffeq_setup}
\alias{diffeq_setup}
\title{Setup diffeqr}
\usage{
diffeq_setup(pkg_check = TRUE, ...)
}
\arguments{
\item{pkg_check}{logical, check for DifferentialEquations.jl package and install if necessary}

\item{...}{Parameters are passed down to JuliaCall::julia_setup}
}
\description{
This function initializes Julia and the DifferentialEquations.jl package.
The first time will be long since it includes precompilation.
Additionally, this will install Julia and the required packages
if they are missing.
}
\examples{

\donttest{ ## diffeq_setup() is time-consuming and requires Julia+DifferentialEquations.jl

diffeqr::diffeq_setup()

}

}
