% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeqr.R
\name{diffeqgpu_setup}
\alias{diffeqgpu_setup}
\title{Setup DiffEqGPU}
\usage{
diffeqgpu_setup()
}
\description{
This function initializes the DiffEqGPU package for GPU-parallelized ensembles.
The first time will be long since it includes precompilation.
}
\examples{

\dontrun{ ## diffeq_setup() is time-consuming and requires Julia+DifferentialEquations.jl

degpu <- diffeqr::diffeqgpu_setup()

}

}
