% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeqr.R
\name{jitoptimize_sde}
\alias{jitoptimize_sde}
\title{Jit Optimize an SDEProblem}
\usage{
jitoptimize_sde(de, prob)
}
\arguments{
\item{de}{the current diffeqr environment}

\item{prob}{an SDEProblem}
}
\description{
This function JIT Optimizes and SDEProblem utilizing the Julia ModelingToolkit
and JIT compiler.
}
\examples{

\donttest{ ## diffeq_setup() is time-consuming and requires Julia+DifferentialEquations.jl

diffeqr::diffeq_setup()

}

}
