% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeqr.R
\name{sde.solve}
\alias{sde.solve}
\title{Solve Stochastic Differential Equations (SDE)}
\usage{
sde.solve(f, g, u0, tspan, p = NULL, alg = "nothing", noise.dims = NULL,
  reltol = 0.01, abstol = 0.01, saveat = NULL)
}
\arguments{
\item{f}{the drift function.}

\item{g}{the diffusion function.}

\item{u0}{the initial condition. Can be a number or (arbitrary dimension) array.}

\item{tspan}{the timespan to solve over. Should be a list of two values: (initial time, end time).}

\item{p}{the parameters. Defaults to no parameters. Can be a number or an array.}

\item{alg}{the algorithm used to solve the differential equation. Defaults to an adaptive choice.
Algorithm choices are done through a string which matches the DifferentialEquations.jl form.}

\item{noise.dims}{list of the dimensions for the noise rate term. Defaults to NULL which gives diagonal noise.}

\item{reltol}{the relative tolerance of the ODE solver. Defaults to 1e-3.}

\item{abstol}{the absolute tolerance of the ODE solver. Defaults to 1e-6}

\item{saveat}{the time points to save values at. Should be an array of times. Defaults to automatic.}
}
\value{
sol. Has the sol$t for the time points and sol$u for the values.
}
\description{
Solves an SDE with du=f(u,p,t)dt + g(u,p,t)dW_t, for u(0)=u0 over the tspan
}
