% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeqr.R
\name{dae.solve}
\alias{dae.solve}
\title{Solve Differential-Algebraic Equations (DAE)}
\usage{
dae.solve(f, du0, u0, tspan, p = NULL, alg = "nothing", reltol = 0.001,
  abstol = 1e-06, saveat = NULL, differential_vars = NULL)
}
\arguments{
\item{f}{the implicit ODE function.}

\item{du0}{the initial derivative. Can be a number or (arbitrary dimension) array.}

\item{u0}{the initial condition. Can be a number or (arbitrary dimension) array.}

\item{tspan}{the timespan to solve over. Should be a list of two values: (initial time, end time).}

\item{p}{the parameters. Defaults to no parameters. Can be a number or an array.}

\item{alg}{the algorithm used to solve the differential equation. Defaults to an adaptive choice.
Algorithm choices are done through a string which matches the DifferentialEquations.jl form.}

\item{reltol}{the relative tolerance of the ODE solver. Defaults to 1e-3.}

\item{abstol}{the absolute tolerance of the ODE solver. Defaults to 1e-6}

\item{saveat}{the time points to save values at. Should be an array of times. Defaults to automatic.}

\item{differential_vars}{boolean array declaring which variables are differential. All falses correspond to
purely algebraic variables.}
}
\value{
sol. Has the sol$t for the time points and sol$u for the values.
}
\description{
Solves a DAE with f(du,u,p,t)=0 for u(0)=u0 over the tspan
}
