% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeqr.R
\name{dde.solve}
\alias{dde.solve}
\title{Solve Delay Differential Equations (DDE)}
\usage{
dde.solve(f, u0, h, tspan, p = NULL, alg = "nothing", reltol = 0.001,
  abstol = 1e-06, saveat = NULL, constant_lags = NULL)
}
\arguments{
\item{f}{the implicit ODE function.}

\item{u0}{the initial condition. Can be a number or (arbitrary dimension) array.}

\item{tspan}{the timespan to solve over. Should be a list of two values: (initial time, end time).}

\item{p}{the parameters. Defaults to no parameters. Can be a number or an array.}

\item{reltol}{the relative tolerance of the ODE solver. Defaults to 1e-3.}

\item{abstol}{the absolute tolerance of the ODE solver. Defaults to 1e-6}

\item{saveat}{the time points to save values at. Should be an array of times. Defaults to automatic.}

\item{du0}{the initial derivative. Can be a number or (arbitrary dimension) array.}
}
\value{
sol. Has the sol$t for the time points and sol$u for the values.
}
\description{
Solves a DDE with f(u,p,t)=0 for u(0)=u0 over the tspan
}
