% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeqr.R
\name{diffeq_setup}
\alias{diffeq_setup}
\title{Setup diffeqr}
\usage{
diffeq_setup(...)
}
\arguments{
\item{Parameters}{are passed down to JuliaCall::julia_setup}
}
\description{
This function initializes Julia and the DifferentialEquations.jl package.
The first time will be long since it includes precompilation.
}
